Give ghost nodes a proper decay lifecycle + legendary tombstones
Dead nodes used to just sit in occupied forever, rendering as a
static '·' dot with no gameplay function. Give them a real
decay arc:

Three-stage visual decay driven by the death_echo countdown
(bumped from u8/60 to u16/600, plus a new GHOST_FADE_TICKS=200
midpoint):

  death_echo > 200  → old role glyph, dim ghost color
  death_echo 1..200 → faint dim '·'
  death_echo == 0   → empty cell (space, no glyph)

The extended window (600 ticks ≈ 30s at 50ms/tick) gives a
visibly fading trace instead of a hard transition. Pass 2 still
dims further, and pass 3 disappears entirely.

Mechanical consequence: at death_echo==0, a ghost cleanup pass
(new, at the end of advance_dying) releases the cell from
self.occupied so new spawns and routing can reclaim the space.
Before this change dead cells were permanent obstacles.

Legendary dead nodes get a carve-out: their cell stays occupied
forever, and they render as a permanent bold '†' tombstone in
the accent color. The inspector still shows the legendary bio,
and a new '✦ legend ✦ {name} falls @ ip (F{N})' mythic log line
fires at the moment of death so named-character deaths hit the
log stream loudly. The map accumulates visible history as
legendary nodes fall — effectively the Round 2 'Lore Tablets'
idea rolled into the ghost decay pass.

+++ src/world/types.rs
@@ -229,7 +229,7 @@ pub struct Node {
     /// at its last position in a dim color so the corpse is visible as
     /// a fading trace before the cell clears. Set when the node first
     /// transitions into `State::Dead`; decays in advance_role_cooldowns.
-    pub death_echo: u8,
+    pub death_echo: u16,
     /// Number of nodes spawned directly from this node as their
     /// parent. Feeds the legendary-node promotion rule along with
     /// the node's age — long-lived, heavily-reproductive nodes get
